commited

+++ libft/ft_itoa.c
@@ -6,7 +6,7 @@
 /*   By: lbarrene <lbarrene@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/16 16:40:53 by lbarrene          #+#    #+#             */
-/*   Updated: 2022/09/18 20:26:34 by lbarrene         ###   ########.fr       */
+/*   Updated: 2022/10/07 17:16:19 by lbarrene         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,8 +30,8 @@ static int	ft_intcheck(int n)
 char	*ft_itoa(int n)
 {
 	char	*str;
-	int		i;
-	int		digit;
+	long	i;
+	long		digit;
 
 	if (n == -2147483648)
 		return (ft_strdup("-2147483648"));
